% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroups.R
\name{a_mmrm_subgroups}
\alias{a_mmrm_subgroups}
\title{Formatted Analysis Function for MMRM Subgroups}
\usage{
a_mmrm_subgroups(.formats)
}
\arguments{
\item{.formats}{(named \code{list})\cr containing the formats for the statistics
to use in the MMRM subgroups table.}
}
\value{
List of formatted analysis functions to be used in
\code{\link[=tabulate_mmrm_subgroups]{tabulate_mmrm_subgroups()}}.
}
\description{
Formatted Analysis Function for MMRM Subgroups
}
\keyword{internal}
