% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmrm.R
\name{fit_mmrm}
\alias{fit_mmrm}
\title{\code{MMRM} Analysis}
\usage{
fit_mmrm(
  vars = list(response = "AVAL", covariates = c(), id = "USUBJID", arm = "ARM", visit =
    "AVISIT"),
  data,
  conf_level = 0.95,
  cor_struct = "unstructured",
  weights_emmeans = "proportional",
  averages_emmeans = list(),
  optimizer = "automatic",
  parallel = FALSE
)
}
\arguments{
\item{vars}{(named \code{list} of \code{string} or \code{character})\cr specifying the variables in the \code{MMRM}.
The following elements need to be included as character vectors and match corresponding columns
in \code{data}:
\describe{
\item{response}{the response variable}
\item{covariates}{the additional covariate terms (might also include interactions)}
\item{id}{the subject ID variable}
\item{arm}{the treatment group variable (factor)}
\item{visit}{the visit variable (factor)}
}
Note that the main effects and interaction of \code{arm} and \code{visit} are by default included in the model.}

\item{data}{a \code{data.frame} with all the variables specified in
\code{vars}. Records with missing values in any independent variables
will be excluded.}

\item{conf_level}{(\code{proportion})\cr confidence level of the interval.}

\item{cor_struct}{a string specifying the covariance structure, defaults to
\code{"unstructured"}. See the details.}

\item{weights_emmeans}{argument from \code{\link[emmeans:emmeans]{emmeans::emmeans()}}, "proportional" by default.}

\item{averages_emmeans}{(\code{list})\cr optional named list of visit levels which should be averaged
and reported along side the single visits.}

\item{optimizer}{a string specifying the optimization algorithm which should be used. By default, "automatic"
will (if necessary) try all possible optimization algorithms and choose the best result. If another algorithm
is chosen and does not give a valid result, an error will occur.}

\item{parallel}{flag that controls whether "automatic" optimizer search can use available free cores on the
machine (not default).}
}
\value{
A \code{tern_mmrm} object which is a list with MMRM results:
\describe{
\item{fit}{The \code{mmrm} object which was fitted to the data. Note that the attribute \code{optimizer}
contains the finally used optimization algorithm, which can be useful for refitting the model later on.}
\item{cov_estimate}{The matrix with the covariance matrix estimate.}
\item{diagnostics}{A list with model diagnostic statistics (REML criterion, AIC, corrected AIC, BIC).}
\item{lsmeans}{This is a list with data frames \code{estimates} and \code{contrasts}.}
\item{vars}{The variable list.}
\item{labels}{Corresponding list with variable labels extracted from \code{data}.}
\item{ref_level}{The reference level for the arm variable, which is always the first level.}
\item{conf_level}{The confidence level which was used to construct the confidence intervals.}
}
}
\description{
Does the \code{MMRM} analysis. Multiple other functions can be called on the result to produce
tables and graphs.
}
\details{
Only Satterthwaite adjusted degrees of freedom (d.f.) are supported, because they
match the results obtained in SAS (confirmed for unstructured and compound symmetry correlation structures).

For the covariance structure (\code{cor_struct}), the user can choose among the following options.
\describe{
\item{unstructured}{Unstructured covariance matrix. This is the most flexible choice and default.
If there are \code{T} visits, then \code{T * (T+1) / 2} variance parameters are used.}
\item{toeplitz}{Heterogeneous Toeplitz covariance matrix,
which uses \code{2 * T - 1} variance parameters.}
\item{ad}{Heterogeneous Ante-Dependence covariance matrix,
which uses \code{2 * T - 1} variance parameters.}
\item{ar1h}{Heterogeneous Auto-Regressive (order 1) covariance matrix,
which uses \code{T + 1} variance parameters.}
\item{ar1}{Homogeneous Auto-Regressive (order 1) covariance matrix,
which uses 2 variance parameters.}
}

For the \code{optimizer}, the user can choose among alternatives to the recommended "automatic",
please see \code{?mmrm::refit_multiple_optimizers} for details. Usually it should not be necessary
to use a specific optimizer and the "automatic" setting should be kept.
}
\examples{
library(dplyr)
library(rtables)

mmrm_results <- fit_mmrm(
  vars = list(
    response = "FEV1",
    covariates = c("RACE", "SEX"),
    id = "USUBJID",
    arm = "ARMCD",
    visit = "AVISIT"
  ),
  data = mmrm_test_data,
  cor_struct = "unstructured",
  weights_emmeans = "equal",
  averages_emmeans = list(
    "VIS1+2" = c("VIS1", "VIS2")
  )
)

}
