% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_plot.R
\name{g_covariance}
\alias{g_covariance}
\title{Function to graph covariance}
\usage{
g_covariance(
  vcov_matrix,
  string = NULL,
  x_var = c("lag", "time_point_distribution"),
  ylab = "",
  xlab = NULL,
  col = tre_col,
  pch = NULL,
  lty = NULL,
  cex = 2,
  legend_position = c("topright", "topleft", "bottomright", "bottomleft"),
  ...
)
}
\arguments{
\item{vcov_matrix}{(\code{matrix})\cr covariance matrix.}

\item{x_var}{(\code{string})\cr can be "lag" or "time_point_distribution" for lag and time distance
respectively.}

\item{legend_position}{(\code{string})\cr denoting where the legend should be shown.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Plot of covariance or correlation structures as a function of lag or time. The covariance structure
needs to be vectorized and lag or time distances computed
}
\examples{
vcov_matrix <- matrix(
c(49, 12, 12, 23),
nrow = 2, ncol = 2,
dimnames = list(
 c(1, 2),
 c(1, 2)
 )
)
g_covariance(vcov_matrix, x_var = "time_point_distribution")
}
