% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mmrm.R
\name{g_mmrm}
\alias{g_mmrm}
\title{\code{MMRM} Plots}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This summarizes available plotting functions for \code{MMRM}.
}
\seealso{
\link{g_mmrm_diagnostic}, \link{g_mmrm_lsmeans}
}
