% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmrm.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Linear Mixed Model Diagnostics}
\usage{
get_diagnostics(fit)
}
\arguments{
\item{fit}{(\code{mmrm})\cr object fit with \code{\link[mmrm:mmrm]{mmrm::mmrm()}}.}
}
\value{
A list with the REML criterion, the AIC, AICc and BIC.
}
\description{
Compute the model diagnostic statistics for a linear mixed model fit.
}
\keyword{internal}
