% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_data.R
\name{h_assert_data}
\alias{h_assert_data}
\alias{h_assert_one_rec_pt_visit}
\alias{h_assert_rsp_var}
\alias{h_assert_visit_var}
\alias{h_assert_id_var}
\alias{h_assert_weights_var}
\title{Assertions for Datasets}
\usage{
h_assert_one_rec_pt_visit(vars, data)

h_assert_rsp_var(vars, data)

h_assert_visit_var(vars, data)

h_assert_id_var(vars, data)

h_assert_weights_var(vars, data)

h_assert_data(vars, data)
}
\arguments{
\item{vars}{(\code{list})\cr variables to use.}

\item{data}{(\code{data.frame})\cr data to use.}
}
\value{
Nothing, only fails with an error if assertion does not hold.
}
\description{
We provide assertion functions to check the input dataset.
}
\section{Functions}{
\itemize{
\item \code{h_assert_one_rec_pt_visit()}: asserts single record per patient and visit.

\item \code{h_assert_rsp_var()}: assert numeric response variable.

\item \code{h_assert_visit_var()}: assert factor visit variable.

\item \code{h_assert_id_var()}: assert subject ID variable.

\item \code{h_assert_weights_var()}: assert numeric weights variable.

\item \code{h_assert_data()}: high-level assertion function to check the dataset.

}}
\keyword{internal}
