% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{h_get_average_visit_specs}
\alias{h_get_average_visit_specs}
\title{Helper Constructing Average of Visits Specifications}
\usage{
h_get_average_visit_specs(emmeans_res, vars, averages)
}
\arguments{
\item{averages}{}
}
\value{
List with \code{coefs} list and \code{grid} data frame.
}
\description{
Helper Constructing Average of Visits Specifications
}
