% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_get_covariate_parts}
\alias{h_get_covariate_parts}
\title{Extraction of Covariate Parts from Character Vector}
\usage{
h_get_covariate_parts(covariates)
}
\arguments{
\item{covariates}{(\code{character})\cr specification in the usual way, see examples.}
}
\value{
Character vector of the covariates involved in \code{covariates} specification.
}
\description{
Extraction of Covariate Parts from Character Vector
}
\keyword{internal}
