% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmrm.R
\name{h_get_diagnostics}
\alias{h_get_diagnostics}
\title{Linear Mixed Model Diagnostics}
\usage{
h_get_diagnostics(fit)
}
\arguments{
\item{fit}{(\code{mmrm})\cr object fit with \code{\link[mmrm:mmrm]{mmrm::mmrm()}}.}
}
\value{
A list with the \code{REML} criterion, the \code{AIC}, \code{AICc} and \code{BIC}.
}
\description{
Compute the model diagnostic statistics for a linear mixed model fit.
}
\keyword{internal}
