% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{h_get_emmeans_res}
\alias{h_get_emmeans_res}
\title{Helper Evaluating \code{emmeans} Results}
\usage{
h_get_emmeans_res(fit, data_complete, vars, weights)
}
\arguments{
\item{weights}{}
}
\value{
List with:
\itemize{
\item \code{object}: \code{emmGrid} object containing \code{emmeans} results.
\item \code{grid}: \code{data.frame} containing the potential arm and the visit variables
together with the sample size \code{n} for each combination.
}
}
\description{
Helper Evaluating \code{emmeans} Results
}
