% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{h_get_relative_reduc_df}
\alias{h_get_relative_reduc_df}
\title{Helper Constructing Data Frame with Relative Reduction vs. Reference Arm}
\usage{
h_get_relative_reduc_df(estimates, vars)
}
\arguments{
\item{vars}{}
}
\value{
A \code{data.frame} with visit, arm and \code{relative_reduc}, the relative
reduction.
}
\description{
Helper Constructing Data Frame with Relative Reduction vs. Reference Arm
}
