% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{h_get_single_visit_estimates}
\alias{h_get_single_visit_estimates}
\title{Helper Calculating Single Visit Least Square Means Results}
\usage{
h_get_single_visit_estimates(emmeans_res, conf_level)
}
\arguments{
\item{conf_level}{}
}
\value{
The \code{data.frame} with the estimates.
}
\description{
Helper Calculating Single Visit Least Square Means Results
}
