% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{h_get_spec_visit_estimates}
\alias{h_get_spec_visit_estimates}
\title{Helper Estimating General Combinations Least Square Means}
\usage{
h_get_spec_visit_estimates(emmeans_res, specs, conf_level, tests = FALSE)
}
\arguments{
\item{conf_level}{}
}
\value{
A \code{data.frame} with estimates.
}
\description{
Helper Estimating General Combinations Least Square Means
}
