% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{h_single_visit_contrast_specs}
\alias{h_single_visit_contrast_specs}
\title{Helper Constructing Single Visit Contrast Specification List}
\usage{
h_single_visit_contrast_specs(emmeans_res, vars)
}
\arguments{
\item{vars}{}
}
\value{
A list with \code{coefs} and \code{grid}, similar as from
\code{\link[=h_get_average_visit_specs]{h_get_average_visit_specs()}}.
}
\description{
Helper Constructing Single Visit Contrast Specification List
}
