% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_plot.R
\name{h_vectorization}
\alias{h_vectorization}
\title{Vectorization helper for Covariance Plot}
\usage{
h_vectorization(vcov_matrix, string = NULL)
}
\arguments{
\item{vcov_matrix}{(\code{matrix})\cr symmetric covariance matrix with identical
row and column names.}

\item{string}{(\code{string})\cr string in the names of \code{vcov_matrix} that precedes
the time point value, see details.}
}
\value{
A data frame with the upper-diagonal elements or a covariance or correlation
matrix. In the context of repeated measures, this matrix contains the association between pairs
of measurements taken at different time points. It contains the following columns: a column
with the upper-diagonal elements of vcov_matrix, the difference between column time
and row time, the lag and the row and column ranks.
}
\description{
This function vectorizes the upper-diagonal elements of a symmetric matrix (e.g. the covariance matrix)
and obtains the lag and time distance between pairs of observations if the time values are part of
the names or are part of the matrix column/row names.
}
\keyword{internal}
