% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{h_is_specified}
\alias{h_is_specified_and_in_data}
\alias{h_check_and_get_label}
\alias{h_labels}
\title{Adding Labels To Variables For Model}
\usage{
h_is_specified(x, vars)

h_is_specified_and_in_data(x, vars, data)

h_check_and_get_label(x, vars, data)

h_labels(vars, data)
}
\arguments{
\item{x}{(\code{character})\cr an element in vars.}

\item{vars}{(\code{list})\cr variables to use.}

\item{data}{(\code{data.frame})\cr data to use.}
}
\description{
Adding Labels To Variables For Model
}
\section{Functions}{
\itemize{
\item \code{h_is_specified()}: checks if element in \code{vars} is not \code{NULL} and not empty.

\item \code{h_is_specified_and_in_data()}: checks if element in vars is not NULL and exists in dataset.

\item \code{h_check_and_get_label()}: gets label for each element in vars.

\item \code{h_labels()}: returns the list of variables with labels.

}}
\keyword{internal}
