% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.mmrm-package.R
\docType{data}
\name{mmrm_test_data}
\alias{mmrm_test_data}
\title{Example dataset for \code{tern.mmrm} package.}
\format{
A \code{tibble} with 800 rows and 7 variables:
\describe{
\item{\code{USUBJID}}{price, in US dollars}
\item{\code{AVISIT}}{visit number}
\item{\code{ARMCD}}{treatment, \code{TRT} or \code{PBO}}
\item{\code{RACE}}{3-category race}
\item{\code{SEX}}{sex}
\item{\code{FEV1_BL}}{FEV1 at baseline (\%)}
\item{\code{FEV1}}{FEV1 at study visits}
}
}
\usage{
mmrm_test_data
}
\description{
Measurements of FEV1 (forced expired volume in one second) is a measure of how quickly the lungs can be emptied.
Low levels of FEV1 may indicate chronic obstructive pulmonary disease (COPD).
}
\keyword{datasets}
