% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_mmrm.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.rtable}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:tidy]{tidy()}}}

  \item{tern}{\code{\link[tern:as.rtable]{as.rtable()}}}
}}

