% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.mmrm-package.R
\docType{package}
\name{tern.mmrm-package}
\alias{tern.mmrm}
\alias{tern.mmrm-package}
\title{\code{tern.mmrm} Package}
\description{
\code{tern.mmrm} is an analysis package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern.mmrm}
  \item \url{https://insightsengineering.github.io/tern.mmrm/}
  \item Report bugs at \url{https://github.com/insightsengineering/tern.mmrm/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})
  \item Daniel Sabanés Bové \email{daniel@rconis.com}
  \item Godwin Yung \email{yungg@gene.com}
  \item Francois Collin
  \item Julia Dedic \email{julia.dedic@roche.com}
  \item Jana Stoilova \email{jana.stoilova@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
