% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_plot.R
\name{h_get_timepoint_vars}
\alias{h_get_timepoint_vars}
\title{Helpers for Covariance Plot}
\usage{
h_get_timepoint_vars(vcov_matrix, string = NULL)
}
\arguments{
\item{vcov_matrix}{(\code{matrix})\cr symmetric covariance matrix with identical
row and column names.}

\item{string}{(\code{string})\cr string in the names of \code{vcov_matrix} that precedes
the time point value, see details.}
}
\value{
This function returns a list with of two sets of numbers:
\code{row_time} and \code{col_time}, identifying the timepoints of the upper triangular
part of \code{vcov_matrix}.
}
\description{
Get the inputted symmetric matrices row and column labels as numeric time points.
}
\details{
The default \code{string} value is \code{NULL}, which assumes that the names
of the input matrix don't have any character string other than time point
value. If a \code{string} is specified, this value should appear in the names
of \code{vcov_matrix}.
}
\examples{
vcov_matrix <- matrix(
  c(49, 24, 12, 23, 24, 35, 11, 20, 12, 11, 24, 14, 23, 20, 14, 107),
  nrow = 4, ncol = 4,
  dimnames = list(
    c("VIS1", "VIS2", "VIS3", "VIS4"),
    c("VIS1", "VIS2", "VIS3", "VIS4")
  )
)
h_get_timepoint_vars(vcov_matrix, string = "VIS")
}
\keyword{internal}
