% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.mmrm-package.R
\docType{data}
\name{mmrm_test_data}
\alias{mmrm_test_data}
\title{Example dataset for \code{tern.mmrm} package.}
\format{
A \code{tibble} with 800 rows and 7 variables:
\itemize{
\item \code{USUBJID}: unique subject identifier.
\item \code{AVISIT}: visit number.
\item \code{ARMCD}: treatment, \code{TRT} or \code{PBO}.
\item \code{RACE}: 3-category race.
\item \code{SEX}: sex.
\item \code{FEV1_BL}: \code{FEV1} at baseline (\%).
\item \code{FEV1}: \code{FEV1} at study visits.
}
}
\usage{
mmrm_test_data
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Measurements of \code{FEV1} (forced expired volume in one second) is a measure of how quickly the lungs can be emptied.
Low levels of \code{FEV1} may indicate chronic obstructive pulmonary disease (\code{COPD}).
}
\keyword{datasets}
