% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tern.mmrm-package}
\alias{tern.mmrm}
\alias{tern.mmrm-package}
\title{tern.mmrm Package}
\description{
tern.mmrm is an analysis package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/nesttemplate}
  \item \url{https://insightsengineering.github.io/nesttemplate/}
  \item Report bugs at \url{https://github.com/insightsengineering/nesttemplate/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

}
