% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.mmrm-package.R
\docType{package}
\name{tern.mmrm-package}
\alias{tern.mmrm}
\alias{tern.mmrm-package}
\title{\code{tern.mmrm} Package}
\description{
\code{tern.mmrm} is an analysis package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern.mmrm}
  \item \url{https://insightsengineering.github.io/tern.mmrm/}
  \item Report bugs at \url{https://github.com/insightsengineering/tern.mmrm/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

}
\keyword{internal}
